[7.1.r1] SoMC Yoshino (plus generic) fixups - platform now bootable

+++ drivers/gpu/msm_legacy/kgsl_trace.h
@@ -1200,7 +1200,7 @@ TRACE_EVENT(kgsl_clock_throttling,
 		int crc_50pct,
 		int crc_more50pct,
 		int crc_less50pct,
-		int adj
+		int64_t adj
 	),
 	TP_ARGS(
 		idle_10pct,
@@ -1214,7 +1214,7 @@ TRACE_EVENT(kgsl_clock_throttling,
 		__field(int, crc_50pct)
 		__field(int, crc_more50pct)
 		__field(int, crc_less50pct)
-		__field(int, adj)
+		__field(int64_t, adj)
 	),
 	TP_fast_assign(
 		__entry->idle_10pct = idle_10pct;
